Updated gubernator-cli with TLS support
