+++ cli-tls.conf
@@ -1,0 +1,5 @@
+GUBER_DEBUG=true
+GUBER_GRPC_ADDRESS=localhost:9081
+GUBER_TLS_CA=certs/ca.pem
+GUBER_TLS_KEY=certs/gubernator.key
+GUBER_TLS_CERT=certs/gubernator.pem
