+++ docker-compose.yaml
@@ -10,8 +10,6 @@ services:
       - GUBER_HTTP_ADDRESS=0.0.0.0:80
       # The address that is advertised to other peers
       - GUBER_ADVERTISE_ADDRESS=gubernator-1:81
-      # Max size of the cache; The cache size will never grow beyond this size.
-      - GUBER_CACHE_SIZE=50000
       # A comma separated list of known gubernator nodes
       - GUBER_MEMBERLIST_KNOWN_NODES=gubernator-1
       #- GUBER_DATA_CENTER=us-east-1
@@ -29,8 +27,6 @@ services:
       - GUBER_HTTP_ADDRESS=0.0.0.0:80
       # The address that is advertised to other peers
       - GUBER_ADVERTISE_ADDRESS=gubernator-2:81
-      # Max size of the cache; The cache size will never grow beyond this size.
-      - GUBER_CACHE_SIZE=50000
       # A comma separated list of known gubernator nodes
       - GUBER_MEMBERLIST_KNOWN_NODES=gubernator-1
       #- GUBER_DATA_CENTER=us-east-1
@@ -48,8 +44,6 @@ services:
       - GUBER_HTTP_ADDRESS=0.0.0.0:80
       # The address that is advertised to other peers
       - GUBER_ADVERTISE_ADDRESS=gubernator-3:81
-      # Max size of the cache; The cache size will never grow beyond this size.
-      - GUBER_CACHE_SIZE=50000
       # A comma separated list of known gubernator nodes
       - GUBER_MEMBERLIST_KNOWN_NODES=gubernator-1
       #- GUBER_DATA_CENTER=us-west-2
@@ -68,10 +62,8 @@ services:
       - GUBER_HTTP_ADDRESS=0.0.0.0:80
       # The address that is advertised to other peers
       - GUBER_ADVERTISE_ADDRESS=gubernator-4:81
-      # Max size of the cache; The cache size will never grow beyond this size.
-      - GUBER_CACHE_SIZE=50000
       # A Comma separated list of known gubernator nodes
-      - GUBER_MEMBERLIST_KNOWN_NODES=gubernator-1
+      - GUBER_MEMBERLIST_KNOWN_NODES=gubernator-1,gubernator-2
       #- GUBER_DATA_CENTER=us-west-2
     ports:
       - "9381:81"
